CORDA-847 - RPC Clent lib refactoring (#3052)
Move Kryo into it's own sub module
