+++ docs/source/changelog.rst
@@ -7,6 +7,8 @@ release, see :doc:`upgrade-notes`.
 Unreleased
 ==========
 
+* Refactor RPC Client Kryo observable serialiser into it's own sub module
+
 * Fix CORDA-1403 where a property of a class that implemented a generic interface could not be deserialised in
   a factory without a serialiser as the subtype check for the class instance failed. Fix is to compare the raw
   type.
